requirements.txt
# Core Framework
gradio

# LangChain/LangGraph
langchain==1.0.8
langgraph==1.0.3
langchain-core==1.0.7
langchain-ollama==1.0.0
langchain-community

# Geospatial
folium==0.20.0
geopandas==1.1.1
shapely==2.1.2
pyproj==3.7.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Environment
python-dotenv==1.0.1

# Additional utilities
requests==2.32.3
pydantic-settings==2.7.2
pydantic==2.7.2

